put all the time getting code in one place
