+++ src/gettime.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+
+uint64_t nowNanos(void);
